Time: 6 ms (75.24%), Space: 22.8 MB (97.73%) - LeetHub

+++ 0875-koko-eating-bananas/0875-koko-eating-bananas.cpp
@@ -6,7 +6,7 @@ public:
 
         while(left < right){
             int mid = left + (right - left)/2;
-            long long hours = 0;
+            int hours = 0;
 
             for(auto banana : piles){
                 hours = hours + (banana + mid - 1) / mid;
